Tweaks to the Adapt_It.vcxproj Visual Studio project file to copy the 13 Windows utility exe files to the deverloper's local Unicode Debug and Unicode Release folders for local development work.
Created a new Windows batch file called _CopyWin32ExeUtils2InstallFolders.bat and added it to the win32_utils folder in the repo. This batch file copies the 13 exe files from the same dirrectory to each of the temporary staging directories: "setup Unicode", "setup Unicode - Minimal", and "setup Unicode - No Html Help". These temporary staging directories is where the Inno Setup scripts fetch them from when creating the 3 Windows installers.

Modified the 3 Unicode Inno Setup scripts to incorporate the 13 exe utility files into the Windows installers, installing them to the same user directory where the Adapt It application is installed.

Still to do is create code within each of the 3 Inno Setup scripts to check is any version 3.x.x of Python for Windows exists on the user's machine, and if not, install a version 3.x.x for Windows during the installation of Adapt It itself (like what is done with git for Windows).