Combine tests.yml workflow into build-dpkg.yml
These were almost the same already. The only difference is that the
build-dpkg workflow was creating and a release if the the change is
a tag.

This combined workflow will trigger on both tags and branches. It will
always create and upload artifacts, but it will only upload the release
if the the change is for a tagged commit..